+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/TakeStockPlanDetail.java
@@ -3,6 +3,7 @@ package com.lframework.xingyun.sc.entity;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.lframework.starter.web.core.dto.BaseDto;
 import com.lframework.starter.web.core.entity.BaseEntity;
+import java.math.BigDecimal;
 import lombok.Data;
 
 /**
@@ -36,27 +37,27 @@ public class TakeStockPlanDetail extends BaseEntity implements BaseDto {
   /**
    * 初始库存数量
    */
-  private Integer stockNum;
+  private BigDecimal stockNum;
 
   /**
    * 盘点数量
    */
-  private Integer oriTakeNum;
+  private BigDecimal oriTakeNum;
 
   /**
    * 修改后的盘点数量
    */
-  private Integer takeNum;
+  private BigDecimal takeNum;
 
   /**
    * 出项数量
    */
-  private Integer totalOutNum;
+  private BigDecimal totalOutNum;
 
   /**
    * 入项数量
    */
-  private Integer totalInNum;
+  private BigDecimal totalInNum;
 
   /**
    * 备注
