+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/TakeStockSheetDetail.java
@@ -3,6 +3,7 @@ package com.lframework.xingyun.sc.entity;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.lframework.starter.web.core.dto.BaseDto;
 import com.lframework.starter.web.core.entity.BaseEntity;
+import java.math.BigDecimal;
 import lombok.Data;
 
 /**
@@ -36,7 +37,7 @@ public class TakeStockSheetDetail extends BaseEntity implements BaseDto {
   /**
    * 盘点数量
    */
-  private Integer takeNum;
+  private BigDecimal takeNum;
 
   /**
    * 备注
